asad.java
Write a program to declare a single-dimensional array a[] and a square matrix b[][] of size N, where N > 2 and N < 10. Allow the user to input positive integers into the single dimensional array.

Perform the following tasks on the matrix:

Sort the elements of the single-dimensional array in ascending order using any standard sorting technique and display the sorted elements.
Fill the square matrix b[][] in the following format:
If the array a[] = {5, 2, 8, 1} then, after sorting a[] = {1, 2, 5, 8}
Then, the matrix b[][] would fill as below:
1
2
5
8
1
2
5
1
1
2
1
2
1
1
2
5
1
1
1
1
​
  
2
2
2
1
​
  
5
5
1
2
​
  
8
1
2
5
​
 

Display the filled matrix in the above format.